+++ tests/BlazorServerFunctions.EndToEndTests/Fixtures/E2EFixture.cs
@@ -92,6 +92,10 @@ public sealed class E2EFixture : IDisposable
             .ConfigurePrimaryHttpMessageHandler(() => factory.Server.CreateHandler())
             .ConfigureHttpClient((_, c) => c.BaseAddress = factory.Server.BaseAddress);
 
+        services.AddHttpClient<IExplicitBindingService, ExplicitBindingServiceClient>()
+            .ConfigurePrimaryHttpMessageHandler(() => factory.Server.CreateHandler())
+            .ConfigureHttpClient((_, c) => c.BaseAddress = factory.Server.BaseAddress);
+
         var grpcChannel = GrpcChannel.ForAddress(
             factory.Server.BaseAddress,
             new GrpcChannelOptions { HttpHandler = factory.Server.CreateHandler() });